+++ packages/sdk/src/config/e2e.ts
@@ -10,6 +10,7 @@ export const e2eConfig: EthereumConfig = {
 		v1: toAddress("0x80f32a12cc4c095e2a409b70e5c96e8515e87dea"),
 		v2: toAddress("0x551E4009116d489e3C5a98405A9c4B601D250B58"),
 		openseaV1: ZERO_ADDRESS,
+		wrapper: ZERO_ADDRESS,
 	},
 	transferProxies: {
 		nft: toAddress("0x66611f8d97688a0af08d4337d7846efec6995d58"),