+++ packages/sdk/src/order/contracts/exchange-opensea-v1.ts
@@ -147,40 +147,16 @@ export const OPENSEA_EXCHANGE_ABI: AbiItem[] = [
 	},
 	{
 		"constant": true,
-		"inputs": [
-			{
-				"name": "addr",
-				"type": "address",
-			},
-		],
-		"name": "testCopyAddress",
+		"inputs": [],
+		"name": "DOMAIN_SEPARATOR",
 		"outputs": [
 			{
 				"name": "",
-				"type": "bytes",
+				"type": "bytes32",
 			},
 		],
 		"payable": false,
-		"stateMutability": "pure",
-		"type": "function",
-	},
-	{
-		"constant": true,
-		"inputs": [
-			{
-				"name": "arrToCopy",
-				"type": "bytes",
-			},
-		],
-		"name": "testCopy",
-		"outputs": [
-			{
-				"name": "",
-				"type": "bytes",
-			},
-		],
-		"payable": false,
-		"stateMutability": "pure",
+		"stateMutability": "view",
 		"type": "function",
 	},
 	{
@@ -357,6 +333,15 @@ export const OPENSEA_EXCHANGE_ABI: AbiItem[] = [
 		"type": "function",
 	},
 	{
+		"constant": false,
+		"inputs": [],
+		"name": "incrementNonce",
+		"outputs": [],
+		"payable": false,
+		"stateMutability": "nonpayable",
+		"type": "function",
+	},
+	{
 		"constant": true,
 		"inputs": [
 			{
@@ -466,7 +451,7 @@ export const OPENSEA_EXCHANGE_ABI: AbiItem[] = [
 			},
 		],
 		"payable": false,
-		"stateMutability": "pure",
+		"stateMutability": "view",
 		"type": "function",
 	},
 	{
@@ -646,7 +631,26 @@ export const OPENSEA_EXCHANGE_ABI: AbiItem[] = [
 			},
 		],
 		"payable": false,
-		"stateMutability": "pure",
+		"stateMutability": "view",
+		"type": "function",
+	},
+	{
+		"constant": true,
+		"inputs": [
+			{
+				"name": "",
+				"type": "address",
+			},
+		],
+		"name": "nonces",
+		"outputs": [
+			{
+				"name": "",
+				"type": "uint256",
+			},
+		],
+		"payable": false,
+		"stateMutability": "view",
 		"type": "function",
 	},
 	{
@@ -928,14 +932,14 @@ export const OPENSEA_EXCHANGE_ABI: AbiItem[] = [
 		"constant": true,
 		"inputs": [
 			{
-				"name": "",
+				"name": "hash",
 				"type": "bytes32",
 			},
 		],
 		"name": "approvedOrders",
 		"outputs": [
 			{
-				"name": "",
+				"name": "approved",
 				"type": "bool",
 			},
 		],
@@ -947,11 +951,73 @@ export const OPENSEA_EXCHANGE_ABI: AbiItem[] = [
 		"constant": false,
 		"inputs": [
 			{
-				"name": "_newOwner",
+				"name": "newOwner",
 				"type": "address",
 			},
 		],
 		"name": "transferOwnership",
+		"outputs": [],
+		"payable": false,
+		"stateMutability": "nonpayable",
+		"type": "function",
+	},
+	{
+		"constant": false,
+		"inputs": [
+			{
+				"name": "addrs",
+				"type": "address[7]",
+			},
+			{
+				"name": "uints",
+				"type": "uint256[9]",
+			},
+			{
+				"name": "feeMethod",
+				"type": "uint8",
+			},
+			{
+				"name": "side",
+				"type": "uint8",
+			},
+			{
+				"name": "saleKind",
+				"type": "uint8",
+			},
+			{
+				"name": "howToCall",
+				"type": "uint8",
+			},
+			{
+				"name": "calldata",
+				"type": "bytes",
+			},
+			{
+				"name": "replacementPattern",
+				"type": "bytes",
+			},
+			{
+				"name": "staticExtradata",
+				"type": "bytes",
+			},
+			{
+				"name": "v",
+				"type": "uint8",
+			},
+			{
+				"name": "r",
+				"type": "bytes32",
+			},
+			{
+				"name": "s",
+				"type": "bytes32",
+			},
+			{
+				"name": "nonce",
+				"type": "uint256",
+			},
+		],
+		"name": "cancelOrderWithNonce_",
 		"outputs": [],
 		"payable": false,
 		"stateMutability": "nonpayable",
@@ -1171,6 +1237,23 @@ export const OPENSEA_EXCHANGE_ABI: AbiItem[] = [
 			},
 		],
 		"name": "OrdersMatched",
+		"type": "event",
+	},
+	{
+		"anonymous": false,
+		"inputs": [
+			{
+				"indexed": true,
+				"name": "maker",
+				"type": "address",
+			},
+			{
+				"indexed": false,
+				"name": "newNonce",
+				"type": "uint256",
+			},
+		],
+		"name": "NonceIncremented",
 		"type": "event",
 	},
 	{