+++ packages/sdk/src/order/contracts/proxy-registry-opensea.ts
@@ -203,7 +203,7 @@ export const proxyRegistryAbi: AbiItem[] = [
 		"constant": false,
 		"inputs": [
 			{
-				"name": "_newOwner",
+				"name": "newOwner",
 				"type": "address",
 			},
 		],