+++ packages/sdk/src/order/fill-order/types.ts
@@ -1,13 +1,7 @@
 import type { Address } from "@rarible/types"
-import type { Part } from "@rarible/ethereum-api-client"
+import type { CryptoPunksAssetType, Erc1155AssetType, Erc721AssetType, Part } from "@rarible/ethereum-api-client"
 import type { Action } from "@rarible/action"
-import type { EthereumTransaction } from "@rarible/ethereum-provider"
-import type { EthereumFunctionCall, EthereumSendOptions } from "@rarible/ethereum-provider"
-import type {
-	CryptoPunksAssetType,
-	Erc1155AssetType,
-	Erc721AssetType,
-} from "@rarible/ethereum-api-client"
+import type { EthereumFunctionCall, EthereumSendOptions, EthereumTransaction } from "@rarible/ethereum-provider"
 import type { Erc1155LazyAssetType, Erc721LazyAssetType } from "@rarible/ethereum-api-client/build/models/AssetType"
 import type { SimpleCryptoPunkOrder, SimpleLegacyOrder, SimpleOpenSeaV1Order, SimpleRaribleV2Order } from "../types"
 import type { NftAssetType } from "../check-asset-type"
@@ -28,7 +22,7 @@ export type LegacyOrderFillRequest =
 export type RaribleV2OrderFillRequest =
   CommonFillRequest<SimpleRaribleV2Order> & { payouts?: Part[], originFees?: Part[] }
 export type OpenSeaV1OrderFillRequest =
-  Omit<CommonFillRequest<SimpleOpenSeaV1Order>, "amount">
+  Omit<CommonFillRequest<SimpleOpenSeaV1Order>, "amount"> & { originFees?: Part[] }
 export type CryptoPunksOrderFillRequest = CommonFillRequest<SimpleCryptoPunkOrder>
 
 export type FillOrderRequest =