+++ packages/test-common/CHANGELOG.md
@@ -30,6 +30,17 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [0.12.66-beta.0](https://github.com/rarible/ethereum-sdk/compare/v0.12.65...v0.12.66-beta.0) (2022-04-28)
+
+
+### Bug Fixes
+
+* fill opensea orders with exchangeWrapper contract ([e6e01e4](https://github.com/rarible/ethereum-sdk/commit/e6e01e4e0a76ef383aa4e75dcfc556d300d4cc77))
+
+
+
+
+
 ## [0.12.64](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.63...v0.12.64) (2022-04-17)